+++ CSharpSourceCode/CampaignMechanics/Crafting/EnchanterTownBehavior.cs
@@ -16,7 +16,6 @@ using TaleWorlds.TwoDimension;
 using TOR_Core.CampaignMechanics.Crafting;
 using TOR_Core.CharacterDevelopment;
 using TOR_Core.Extensions;
-using TOR_Core.Items;
 using TOR_Core.Utilities;
 using static TOR_Core.Utilities.TORConstants;
