+++ CSharpSourceCode/CampaignMechanics/Crafting/EnchantingVM.cs
@@ -8,7 +8,6 @@ using TaleWorlds.Library;
 using TaleWorlds.LinQuick;
 using TOR_Core.Extensions;
 using TOR_Core.Items;
-using TOR_Core.Models;
 using TOR_Core.Utilities;
 
 namespace TOR_Core.CampaignMechanics.Crafting
