+++ CSharpSourceCode/CampaignMechanics/Crafting/EnchantmentIngredientLootCampaignBehavior.cs
@@ -8,7 +8,6 @@ using TaleWorlds.CampaignSystem.Settlements;
 using TaleWorlds.Core;
 using TaleWorlds.LinQuick;
 using TOR_Core.Extensions;
-using TOR_Core.Items;
 
 namespace TOR_Core.CampaignMechanics.Crafting;
 /// <summary>
