+++ CSharpSourceCode/CampaignMechanics/Crafting/TOREnchantmentIngredientsModel.cs
@@ -4,7 +4,6 @@ using TaleWorlds.CampaignSystem.MapEvents;
 using TaleWorlds.Core;
 using TOR_Core.CharacterDevelopment;
 using TOR_Core.Extensions;
-using TOR_Core.Items;
 using TOR_Core.Utilities;
 
 namespace TOR_Core.CampaignMechanics.Crafting;
