+++ CSharpSourceCode/CampaignMechanics/Crafting/TORSmithingModel.cs
@@ -9,7 +9,7 @@ using TOR_Core.CharacterDevelopment;
 using TOR_Core.Extensions;
 using TOR_Core.Utilities;
 
-namespace TOR_Core.Models
+namespace TOR_Core.CampaignMechanics.Crafting
 {
     /// <summary>
     /// Categories for weapon templates used in culture-based order filtering.
@@ -132,7 +132,7 @@ namespace TOR_Core.Models
             return false;
         }
 
-        public override int GetEnergyCostForRefining(ref Crafting.RefiningFormula refineFormula, Hero hero)
+        public override int GetEnergyCostForRefining(ref TaleWorlds.Core.Crafting.RefiningFormula refineFormula, Hero hero)
         {
             var value = base.GetEnergyCostForRefining(ref refineFormula, hero);
             return ApplyEnergyCostModifiers(value, hero);
@@ -170,7 +170,7 @@ namespace TOR_Core.Models
             return value;
         }
 
-        public override IEnumerable<Crafting.RefiningFormula> GetRefiningFormulas(
+        public override IEnumerable<TaleWorlds.Core.Crafting.RefiningFormula> GetRefiningFormulas(
             Hero weaponsmith)
         {
             var values = base.GetRefiningFormulas(weaponsmith);
@@ -179,12 +179,12 @@ namespace TOR_Core.Models
 
             if (weaponsmith.HasCareer(TORCareers.Runelord))
             {
-                var newValues = new List<Crafting.RefiningFormula>();
+                var newValues = new List<TaleWorlds.Core.Crafting.RefiningFormula>();
                 foreach (var value in values)
                 {
                     if (weaponsmith.HasCareerChoice("ForgefireBurningPassive1") && value.Output == CraftingMaterials.Charcoal)
                     {
-                        var entry = new Crafting.RefiningFormula(value.Input1, value.Input1Count, value.Input2, value.Input2Count, value.Output,
+                        var entry = new TaleWorlds.Core.Crafting.RefiningFormula(value.Input1, value.Input1Count, value.Input2, value.Input2Count, value.Output,
                             value.OutputCount + 1);
                         newValues.Add(entry);
                         continue;
@@ -192,7 +192,7 @@ namespace TOR_Core.Models
                     if (weaponsmith.HasCareerChoice("ForgefireBurningPassive2") && value.Output is CraftingMaterials.Iron1 or CraftingMaterials.Iron2 or CraftingMaterials.Iron3 or CraftingMaterials.Iron4)
                     {
 
-                        var entry = new Crafting.RefiningFormula(value.Input1, value.Input1Count, value.Input2, value.Input2Count, value.Output,
+                        var entry = new TaleWorlds.Core.Crafting.RefiningFormula(value.Input1, value.Input1Count, value.Input2, value.Input2Count, value.Output,
                             value.OutputCount * 2);
                         newValues.Add(entry);
                         continue;
