+++ CSharpSourceCode/Items/ItemTrait.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Serialization;
+using TOR_Core.CampaignMechanics.Crafting;
 using TOR_Core.Extensions.ExtendedInfoSystem;
 using static TaleWorlds.Core.ItemObject;
