+++ CSharpSourceCode/Models/TORFaithModel.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.CampaignSystem.Party;
 using TaleWorlds.Core;
+using TOR_Core.CampaignMechanics.Crafting;
 using TOR_Core.CampaignMechanics.Religion;
 using TOR_Core.CharacterDevelopment;
 using TOR_Core.Extensions;
 using TOR_Core.Extensions.ExtendedInfoSystem;
-using TOR_Core.Items;
 using TOR_Core.Utilities;
 
 namespace TOR_Core.Models
