+++ CSharpSourceCode/TOR_Core.csproj
@@ -506,6 +506,9 @@
     <Compile Include="CampaignMechanics\CharacterCreation\TORCharacterCreationContentHandler.cs" />
     <Compile Include="CampaignMechanics\CharacterCreation\TORCharacterCreationException.cs" />
     <Compile Include="CampaignMechanics\Companions\TORCompanionsCampaignBehavior.cs" />
+    <Compile Include="CampaignMechanics\Crafting\CraftingModule.cs" />
+    <Compile Include="CampaignMechanics\Crafting\CraftingPatches.cs" />
+    <Compile Include="CampaignMechanics\Crafting\CraftingVMExtension.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantableItemVM.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantableTraitVM.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchanterTownBehavior.cs" />
@@ -515,12 +518,18 @@
     <Compile Include="CampaignMechanics\Crafting\EnchantingScreen.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantingState.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantingVM.cs" />
+    <Compile Include="CampaignMechanics\Crafting\EnchantmentBlueprintScript.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantmentHelper.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantmentShopHelper.cs" />
     <Compile Include="CampaignMechanics\Crafting\EnchantmentIngredientLootCampaignBehavior.cs" />
     <Compile Include="CampaignMechanics\Crafting\LootCampaignBehavior.cs" />
     <Compile Include="CampaignMechanics\Crafting\PriestBehavior.cs" />
+    <Compile Include="CampaignMechanics\Crafting\RefinementVMExtension.cs" />
+    <Compile Include="CampaignMechanics\Crafting\TorEnchantingIngredients.cs" />
+    <Compile Include="CampaignMechanics\Crafting\TOREnchantmentCraftingModel.cs" />
+    <Compile Include="CampaignMechanics\Crafting\TOREnchantmentIngredientsModel.cs" />
     <Compile Include="CampaignMechanics\Crafting\TORArtisanDistrictCampaignBehavior.cs" />
+    <Compile Include="CampaignMechanics\Crafting\TORSmithingModel.cs" />
     <Compile Include="CampaignMechanics\CustomDialogs\BloodKissSceneNotificationItem.cs" />
     <Compile Include="CampaignMechanics\CustomDialogs\CareerSwitchCampaignBehavior.cs" />
     <Compile Include="CampaignMechanics\CustomDialogs\ConversationTags\AsraiTag.cs" />
@@ -716,12 +725,10 @@
     <Compile Include="Extensions\SkillExtensions.cs" />
     <Compile Include="Extensions\TeamExtensions.cs" />
     <Compile Include="Extensions\UI\BaseViewModelExtension.cs" />
-    <Compile Include="Extensions\UI\CraftingVMExtension.cs" />
     <Compile Include="Extensions\UI\ConversationItemVMExtension.cs" />
     <Compile Include="Extensions\UI\HeroEncyclopediaVMExtension.cs" />
     <Compile Include="Extensions\UI\MissionConversationVMExtension.cs" />
     <Compile Include="Extensions\UI\PartyCharacterVMExtension.cs" />
-    <Compile Include="Extensions\UI\RefinementVMExtension.cs" />
     <Compile Include="Extensions\UI\UnitEncyclopediaVMExtension.cs" />
     <Compile Include="Extensions\UI\ViewModelExtensionAttribute.cs" />
     <Compile Include="Extensions\UI\ViewModelExtensionManager.cs" />
@@ -746,7 +753,6 @@
     <Compile Include="HarmonyPatches\CharacterObjectPatches.cs" />
     <Compile Include="HarmonyPatches\CaravanVisualPatch.cs" />
     <Compile Include="HarmonyPatches\ConversationPatches.cs" />
-    <Compile Include="HarmonyPatches\CraftingPatches.cs" />
     <Compile Include="HarmonyPatches\CustomBattlePatches.cs" />
     <Compile Include="HarmonyPatches\CustomResourcePatches.cs" />
     <Compile Include="HarmonyPatches\CustomWorldMapPatch.cs" />
@@ -791,14 +797,12 @@
     <Compile Include="Items\ExtendedItemObjectProperties.cs" />
     <Compile Include="Items\InventoryUseScriptsCampaignBehavior.cs" />
     <Compile Include="Items\InventoryUseScripts\CustomResourceContainerScript.cs" />
-    <Compile Include="Items\InventoryUseScripts\EnchantmentBlueprintScript.cs" />
     <Compile Include="Items\InventoryUseScripts\IInventoryUseScript.cs" />
     <Compile Include="Items\InventoryUseScripts\SkillBookScript.cs" />
     <Compile Include="Items\InventoryUseScripts\StartInkStoryScript.cs" />
     <Compile Include="Items\ItemTrait.cs" />
     <Compile Include="Items\ItemTraitAgentComponent.cs" />
     <Compile Include="Items\ItemTraitManager.cs" />
-    <Compile Include="Items\TorEnchantingIngredients.cs" />
     <Compile Include="Items\TorImageIdentifierWidget.cs" />
     <Compile Include="Items\TorInventoryItemTupleWidget.cs" />
     <Compile Include="Items\TorItemMenuVM.cs" />
@@ -850,8 +854,6 @@
     <Compile Include="Models\TORAllianceModel.cs" />
     <Compile Include="Models\TORSettlementProsperityModel.cs" />
     <Compile Include="Models\TORTradeAgreementModel.cs" />
-    <Compile Include="Models\TOREnchantmentCraftingModel.cs" />
-    <Compile Include="Models\TOREnchantmentIngredientsModel.cs" />
     <Compile Include="Models\TOREncounterGameMenuModel.cs" />
     <Compile Include="Models\TOREncounterModel.cs" />
     <Compile Include="Models\TOREquipmentSelectionModel.cs" />
@@ -877,7 +879,6 @@
     <Compile Include="Models\TORSettlementLoyaltyModel.cs" />
     <Compile Include="Models\TORSettlementMilitiaModel.cs" />
     <Compile Include="Models\TORSiegeEngineCalculationModel.cs" />
-    <Compile Include="Models\TORSmithingModel.cs" />
     <Compile Include="Models\TORStrikeMagnitudeModel.cs" />
     <Compile Include="Models\TORTournamentModel.cs" />
     <Compile Include="Models\TORTroopSupplierModel.cs" />
